storage/requirements.txt
connexion[flask,uvicorn,swagger-ui]
PyYAML
kafka-python
sqlalchemy
mysqlclient
pymysql
httpx
apscheduler
cryptography